charting working but not customizable
