--- TESTING/static/test.js
@@ -1,13 +1,0 @@
-var ctx = document.getElementById("myChart");
-
-var myChart = new Chart(ctx, {
-  type: 'line',
-  data: {
-    labels: steps,
-    datasets: [
-      { 
-        data: cultures      
-    }
-    ]
-  }
-});
