--- TESTING/templates/simulation.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <!--suppress JSUnresolvedLibraryURL -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.0/jquery.min.js"></script>
-    <!--suppress JSUnresolvedLibraryURL -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/js/bootstrap.min.js"></script>
-    <!--suppress JSUnresolvedLibraryURL -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script>
-
-    <title>
-        Prova
-    </title>
-</head>
-
-<body>
-    <h2 id="headline">params here</h2>
-    <canvas id="myChart"></canvas>
-    <script type="text/javascript">
-        /* var steps = {{ steps }};
-        var cultures = {{ cultures }} */
-    </script>
-    <script src="../static/test2.js"> </script>
-</body>
-
-</html>
