--- static/scripts.js
@@ -1,2 +1,0 @@
-var l = [[[1, 2, 3], [32, 22, 21]], [[123321, 3123123], [13213213, 3123213]]];
-console.log(l[0][1]);
