--- static/style.css
@@ -1,28 +1,0 @@
-h1 {
-    font-family: "Monoton";
-    font-size: 70px;
-    margin: 0.2em;
-}
-
-h3 {
-    color: blueviolet;
-}
-
-table,
-tr,
-td {
-    border: 1px solid black;
-}
-
-table {
-    border-collapse: collapse;
-    margin: 0 auto;
-}
-
-tr {
-    height: 20px;
-}
-
-td {
-    width: 20px;
-}
