+++ static/styles.css
@@ -1,0 +1,4 @@
+.chart-container {
+    width: 500px;
+    height: 600px
+}
