+++ static/test2.js
@@ -45,16 +45,11 @@ $(document).ready(function () {
     };
 
     const context = document.getElementById('myChart').getContext('2d');
-
     const lineChart = new Chart(context, config);
-
     const source = new EventSource("/chart-data");
 
     source.onmessage = function (event) {
         const data = JSON.parse(event.data);
-        console.log("Banzo")
-        change = document.getElementById("headline"); 
-        change.innerText=data.step;
         if (config.data.labels.length === 20) {
             config.data.labels.shift();
             config.data.datasets[0].data.shift();
@@ -63,12 +58,10 @@ $(document).ready(function () {
         config.data.datasets[0].data.push(data.cultures);
         lineChart.update();
         if (data.step === -1000) {
-            console.log("Banzai!!!!!!!")
             change = document.getElementById("headline"); 
-            change.innerText="POLLO";
+            change.innerText="Equilibrium reached";
             lineChart.stop();
             source.close();
         }
     }
-    
 });
