+++ templates/index.html
@@ -1,18 +1,53 @@
-{% extends "layout.html" %}
+<!DOCTYPE html>
+<html lang="en">
 
-{% block main %}
+<head>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.0/jquery.min.js"></script>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/js/bootstrap.min.js"></script>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script>
+    
+    <!-- CSS only -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
+    <link rel="stylesheet" href="/static/styles.css">
+    <title>
+        Prova
+    </title>
+</head>
 
-<h1>Axelrod</h1>
-<h3 id="pollo">POLLLO</h3>
-<h2>Choose parameters</h2>
-<form id="paramPicker" action="/" method="post">
-    Grid size:
-    <input type="number" id="gridSize" name="size" min="2" max="100" value="4">
-    Number of cultures:
-    <input type="number" id="no_features" name="no_features" min="6" , max="6" value="6">
-    Number of traits:
-    <input type="number" id="no_traits" name="no_traits" min="1" , max="16" value="9">
-    <input type="submit">
-</form>
+<body>
+    <h1>Axelrod: modeling cultural dissemination</h1>
+    <h2>Choose parameters</h2>
+    <form id="paramPicker" action="/" method="post">
+        Grid size:
+        <input type="number" id="gridSize" name="size" min="2" max="100" value="4">
+        Number of cultures:
+        <input type="number" id="no_features" name="no_features" min="6" , max="6" value="6">
+        Number of traits:
+        <input type="number" id="no_traits" name="no_traits" min="1" , max="16" value="9">
+        <input type="submit">
+    </form>
+    <div class="container">
+        <div class="row">
+            <div class="col">
+                <h2 id="headline">params here</h2>
+                <div class="chart-container">
+                    <canvas id="myChart"></canvas>
+                </div>
+            </div>
+            <div class="col">
+                <h2 id="headline2">BANZAI!!!</h2>
+                <div class="chart-container">
+                    <canvas id="myChart2"></canvas>
+                </div>
+            </div>  
+        </div>
+    </div>
+    
+    <script src="../static/test2.js"> </script>
+</body>
 
-{% endblock %}
+</html>
