+++ templates/simulation.html
@@ -1,21 +1,27 @@
-{% extends "layout.html" %}
+<!DOCTYPE html>
+<html lang="en">
 
-{% block main %}
+<head>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.0/jquery.min.js"></script>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/js/bootstrap.min.js"></script>
+    <!--suppress JSUnresolvedLibraryURL -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.8.0/Chart.min.js"></script>
 
-<h1>SIMULATION</h1>
-<h3 id="pollo">FAI STA CAZZO DI GRIGLIA MADONNA</h3>
-<h2> DATASETS {{ dataJS | tojson }} </h2>
-<canvas id="myChart"></canvas>
-<h2>SIZE {{ size }}</h2>
-<h3 id = "verifica">""</h3>
-<table id="grid">
-</table>
-<script>
-    var size = "{{ size }}";
-    var lattice ="{{ lattice }}"
-</script>
-<!-- <script src="../static/scripts2.js"> 
-</script>
-<script src="../static/chart.js"></script>-->
+    <title>
+        Prova
+    </title>
+</head>
 
-{% endblock %}
+<body>
+    <h2 id="headline">params here</h2>
+    <canvas id="myChart"></canvas>
+    <script type="text/javascript">
+        /* var steps = {{ steps }};
+        var cultures = {{ cultures }} */
+    </script>
+    <script src="../static/test2.js"> </script>
+</body>
+
+</html>
